docs: fixs docs build
