+++ pandasai/prompts/__init__.py
@@ -1,4 +1,5 @@
-from .base import AbstractPrompt, FileBasedPrompt
+from .base import AbstractPrompt
+from .file_based_prompt import FileBasedPrompt
 from .correct_error_prompt import CorrectErrorPrompt
 from .generate_python_code import GeneratePythonCodePrompt
