+++ pandasai/prompts/clarification_questions_prompt.py
@@ -18,7 +18,7 @@ Json:
 import json
 from typing import List
 import pandas as pd
-from .base import FileBasedPrompt
+from .file_based_prompt import FileBasedPrompt
 
 
 class ClarificationQuestionPrompt(FileBasedPrompt):
